# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.ErrorResponse:
    properties:
      error:
        type: string
      status:
        type: string
    type: object
  api.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  scraper.JobPosting:
    properties:
      company:
        type: string
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      location:
        type: string
      platform_job_id:
        type: string
      source:
        $ref: '#/definitions/scraper.ScraperType'
      summary:
        type: string
      title:
        type: string
      url:
        type: string
    type: object
  scraper.ScraperType:
    enum:
    - indeed
    - linkedin
    type: string
    x-enum-varnames:
    - Indeed
    - LinkedIn
host: localhost:8080
info:
  contact: {}
  description: This is a job scraper application.
  termsOfService: http://swagger.io/terms/
  title: Job Scraper API
  version: "1.0"
paths:
  /jobs:
    get:
      consumes:
      - application/json
      description: Get a list of jobs
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/scraper.JobPosting'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Get jobs
      tags:
      - jobScraper
  /scrape:
    post:
      consumes:
      - application/json
      description: Start scraping jobs based on the provided configuration
      parameters:
      - description: JobPosting Title
        in: query
        name: jobTitle
        required: true
        type: string
      - description: Country
        in: query
        name: country
        required: true
        type: string
      - default: 1
        description: Number of Pages
        in: query
        name: pages
        type: integer
      - description: Source of job listings (indeed or linkedin)
        enum:
        - indeed
        - linkedin
        in: query
        name: source
        required: true
        type: string
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "501":
          description: Not Implemented
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      summary: Start scraping
      tags:
      - jobScraper
swagger: "2.0"